recommendation_app/requirements.txt
streamlit
streamlit-aggrid==0.2.2-2
pandas
numpy
scikit-learn
matplotlib
seaborn
scipy
tensorflow
scikit-surprise==1.1.3